Feat: Completed Bounty_002
