--- py/timings.tsv
@@ -1,8 +1,0 @@
-scene	seconds	status
-bicycle	206.88	OK
-bonsai	328.82	OK
-counter	280.33	OK
-kitchen	303.57	OK
-room	323.88	OK
-stump	143.91	OK
-garden	207.85	OK
